various and related to MARSSParamCIs changes
* MARSSparamCIs gave the wrong s.e. for variances and covariances when
method="hessian".  Also wrong CIs for variances and covariances when
var-cov matrix was non-diagonal. There were a series of issues related
to back-transforming from the hessian of a chol-transformed var-cov
matrix ( Sigma=chol%*%t(chol) ).
-- MARSSparamCIs gets the hessian numerically using a var-cov matrix
that has been transformed with a Cholesky decomposition to ensure it
stays positive-definite.  The upper and lower CIs were computed via
back-transformation but the s.e.'s for the variances and covariances
were not corrected.  So the s.e. for R and Q were wrong in all cases.
-- There was a bug in MARSShessian error in subscripting the d matrix
when doing the chol transform.  Caused NAs for cases with non-diagonal
matrices.
-- The s.e.'s and CIs were completely wrong for non-diagonal matrices
because the elements of the chol transformed matrices do not correspond
one-to-one to the non-transformed matrices.  E.g. the untransformed
[2,2]^2 is chol transformed [1,2]^2+[2,2]^2
-- The hessian used was for the chol-transformation and that MLE is not
necessarily equal to the MLE for the untransformed var-cov matrices.
-- Fix:
#1 Created new function MARSShessian.H1989 which uses the Harvey 1989
recursion to compute the observed Fisher Information matrix. This is the
hessian for the untransformed var-cov parameters.  So CIs on variances
can be negative since the variance of the MLE is being approximated by a
MVN (which can be negative).
#2 This is now the default for method="kem" (Kalman-EM algorithm).
#3 Fixed the back-transformation for s.e.'s when var-cov matrices are
diagonal.  For "unconstrained" matrices, I use simulate from the
$parMean and $parSigma, compute the var-cov matrix, Renamed
MARSShessian.chol to indicate that it is for the Cholesky transformed
var-cov matrices.
#4 MARSShessian.chol is the default for method="BFGS", because BFGS is
finding the MLEs for Cholesky transformed var-cov matrices.