various
* residuals.marssMLE had a bug that lead to erroneous standardized
residuals when Z was non-diagonal (and thus also > 1 row). Also changed
residuals.MARSSMLE so it returns residuals (which would be equal to 0)
when there are 0s on the diagonal of Q or R
* residuals.marssMLE had a bug so it didn't return the residuals at t=1
when x0 was fixed
* residuals.marssMLE had a bug so that it returned the wrong residuals
if there were any missing values.  Fix involved implementing the missing
values modifications for the Kalman filter described in Shumway and
Stoffer.