comment out code that 0s out Vtt when R=0
When Z is diagonal (say) and R=0 and y not missing, Vtt must be 0. I was setting Vtt to 0 in these cases to improve the numerical accuracy. So that numerical errors were not propogated.

But the conditions were not right. I had Z!=0 but it is that colSums Z == 1 AND corresponding Q !=0.  This code is not needed so I commented it out.